% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_test.R
\name{chisq_test_odds}
\alias{chisq_test_odds}
\title{Chi-square Test Sample Size Given Odds Ratio}
\usage{
chisq_test_odds(odds_ratio = NULL, p_control = NULL,
  p_treatment = NULL, power = NULL, conf_level = 0.95,
  sample_ratio = 1, visualize = FALSE)
}
\arguments{
\item{odds_ratio}{The expected odds ratio. That is, the ratio of the odds of
the outcome in the test group relative to the control group. Optional,
but see \emph{Details}.}

\item{p_control}{Your guess for prevalence of outcome in the control group.
Optional but see \strong{Details}.}

\item{p_treatment}{Your guess for prevalence of outcome in the test group.
Optional but see \strong{Details}.}

\item{power}{The ability of the test to detect an effect where there is one.
Power = 1 - Prob(Type 2 error). Optional. See \strong{Value} for details.}

\item{conf_level}{Desired confidence level. Defaults to 95\%.}

\item{sample_ratio}{Ratio of test group to control group. 1 is even split.}

\item{visualize}{Whether to plot power or prevalence of outcome in the
control group vs sample size. Can be used to help make a decision.}
}
\value{
If \code{power} was not provided, returns vector containing
possible power values and the appropriate sample size for each \%.
If all values were provided, returns a single sample size estimate.
}
\description{
Calculates sample size for chi-squared test of independence
given the odds ratio.
}
\details{
The function only needs to know two of the following three:
\code{odds_ratio}, \code{p_control}, and \code{p_treatment}. If given
all three, it will check to make sure the odds ratio is correct. It
will figure out the missing third value from the other two.
}
\section{References}{

Wang, H., Chow, S.-C., & Li, G. (2002). On sample size calculation based on
odds ratio in clinical trials. \emph{Journal of Biopharmaceutical
Statistics}, \strong{12}(4), 471-483.
\href{http://doi.org/10.1081/BIP-120016231}{doi:10.1081/BIP-120016231}
}

\examples{
chisq_test_odds(p_treatment = 0.4, p_control = 0.25, power = 0.8)
chisq_test_odds(odds_ratio = 2, p_control = 0.4, power = c(0.8, 0.9, 0.95))
chisq_test_odds(odds_ratio = 2, p_control = 0.4)
chisq_test_odds(odds_ratio = 2, p_control = 0.4, visualize = TRUE)
}
\seealso{
\code{\link[=chisq_test_effect]{chisq_test_effect()}}
}
\author{
Mikhail Popov
}
