% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hive.R
\name{date_clause}
\alias{date_clause}
\title{Generate a Date Clause for a Hive query}
\usage{
date_clause(date)
}
\arguments{
\item{date}{if \code{NULL}, yesterday will be used}
}
\value{
a list containing two elements: "date_clause" and "date"; the
returning of the date allows you to include it
}
\description{
What it says on the tin; generates a
\code{WHERE year = foo AND month = bar}
that can then be combined with other elements to form a Hive query.
}
\seealso{
\code{\link[=extract_ymd]{extract_ymd()}}
}
