% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exact_binom}
\alias{exact_binom}
\title{Sample size for exact, one sample binomial test}
\usage{
exact_binom(constant_prop, effect_size, alpha = 0.05, power = 0.8,
  two_tail = TRUE)
}
\arguments{
\item{constant_prop}{The proportion under the null hypothesis.}

\item{effect_size}{Positive size of the difference between your null
hypothesis and the alternative hypothesis that you hope to detect.
\strong{Heads-up} that values less than 1\% might take a while to calculate.}

\item{alpha}{Probability of rejecting the null hypothesis even though it is
true.}

\item{power}{Probability of rejecting the null hypothesis (getting a
significant result) when the real difference is equal to the minimum
effect size.}

\item{two_tail}{Whether to perform two-tail or one-tail power analysis.
\code{TRUE} (default) tests in both directions of difference.}
}
\description{
Estimates sample size required to detect difference from a
constant proportion.
}
\examples{
exact_binom(0.75, 0.03)
}
\references{
\href{http://www.biostathandbook.com/power.html}{Power analysis} and
\href{http://www.biostathandbook.com/exactgof.html}{Exact test of goodness-of-fit} from
John H. McDonald's \href{http://www.biostathandbook.com/}{Handbook of Biological Statistics}
}
