% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_ymd}
\alias{extract_ymd}
\title{Extract YMD components from a date}
\usage{
extract_ymd(x)
}
\arguments{
\item{x}{\code{Date} vector}
}
\description{
Extracts a triple of year, month, day from date(s). If a single
date is provided, useful with \code{zeallot} package's multi-assignment
operator. If multiple dates are provided, a \code{tibble} of components (as
columns) are returned, which can be used with \code{purrr}'s multi-column
mappings.
}
\examples{
\dontrun{
library(zeallot)

c(year, month, day) \%<-\% extract_ymd(Sys.Date())
}
}
