% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{FiveThirtyNine}
\alias{FiveThirtyNine}
\alias{theme_fivethirtynine}
\title{Theme inspired by fivethirtyeight.com plots}
\usage{
theme_fivethirtynine(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
A modification of \code{ggthemes::theme_fivethirtyeight}
}
\details{
Basically it adds axis titles (with some modification on the y to
 allow for long titles) back in and does a small amount of reduction of the
 overall plot size to avoid an absolute ton of extraneous spacing.
}

