% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{Formatters}
\alias{Formatters}
\alias{percent2}
\alias{pretty_num}
\title{Formatters}
\usage{
percent2(x, digits = 1, add_plus = FALSE)

pretty_num(x, ...)
}
\arguments{
\item{x}{A vector to format}

\item{...}{Additional parameters to pass to \code{\link[base:prettyNum]{base::prettyNum()}}}
}
\description{
Formatting utilities
}
\details{
\itemize{
\item \code{percent2}: multiply by one hundred, display percent sign, clear "NA%"'s,
and optionally prepend a "+" to positive percentages
\item \code{pretty_num}: shortcut to formatting 1e6 as 1,000,000
}
}
