% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{get_logfile}
\alias{get_logfile}
\title{Retrieve a vector of sampled log files}
\usage{
get_logfile(earliest = NULL, latest = NULL)
}
\arguments{
\item{earliest}{a \code{Date} object. Set to \code{NULL} by default, which triggers
the retrieval of all log file names.}

\item{latest}{a \code{Date} object; set to \code{NULL} by default. In the event that
\code{earliest} is set but \code{latest} is not, the files retrieved will span from
\code{earliest} to the current date; in the event that both arguments are set,
the retrieved files will be those in that range.}
}
\value{
A vector of filenames that can be passed into \code{\link[=read_sampled_log]{read_sampled_log()}}
}
\description{
Grabs sampled log files to be piped into \code{\link[=read_sampled_log]{read_sampled_log()}}.
By default this retrieves all sampled log files; it can be used to
retrieve a particular date range of files through the "earliest" and
"latest" arguments.
}
\author{
Os Keyes
}
