% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\docType{data}
\name{geom_flat_violin}
\alias{geom_flat_violin}
\alias{GeomFlatViolin}
\title{Flat violin plot}
\source{
Gist: \href{https://gist.github.com/dgrtwo/eb7750e74997891d7c20}{dgrtwo/geom_flat_violin.R}
}
\usage{
geom_flat_violin(mapping = NULL, data = NULL, stat = "ydensity",
  position = "dodge", trim = TRUE, scale = "area",
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{trim}{If \code{TRUE} (default), trim the tails of the violins
to the range of the data. If \code{FALSE}, don't trim the tails.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{geom, stat}{Use to override the default connection between
\code{geom_violin} and \code{stat_ydensity}.}
}
\description{
Violin plots are a compact display of continuous distributions
but are usually mirrored to mimick boxplots. The "flat" version removes
that mirrorness and makes the violin plots less...suggestive.
}
\examples{
\dontrun{
ggplot(diamonds, aes(cut, carat)) +
  geom_flat_violin() +
  coord_flip()
}
}
\author{
\href{https://github.com/dgrtwo}{David Robinson}
}
\keyword{datasets}
