% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{global_query}
\alias{global_query}
\title{Global SQL queries for analytics-store.eqiad.wmnet}
\usage{
global_query(query, project_type = "all")
}
\arguments{
\item{query}{the SQL query you want to run}

\item{project_type}{what class of wiki (e.g. "wikisource", "wiktionary")
you want to run against. Set to "all" by default.}
}
\description{
\code{global_query} is a simple wrapper around the MySQL queries
that allows a useR to send a query to all production dbs on
analytics-store.eqiad.wmnet, joining the results from each query into a
single object.
}
\seealso{
\code{\link[=mysql_read]{mysql_read()}} for querying an individual db, \code{\link[=from_mediawiki]{from_mediawiki()}}
for converting MediaWiki timestamps into \code{POSIXlt} timestamps, or
\code{\link[=query_hive]{query_hive()}} for accessing the Hive datastore
}
\author{
Oliver Keyes
}
