% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/interleaved.R
\docType{data}
\name{interleaved_bootstraps}
\alias{interleaved_bootstraps}
\alias{interleaved_confint}
\alias{interleaved_preference}
\alias{interleaved}
\alias{interleaved_data}
\alias{interleaved_data_a}
\alias{interleaved_data_b}
\title{Interleaved search results}
\format{\code{interleaved_data*} are \code{data.frame}-s of generated search sessions with
the following columns:
\describe{
\item{session_id}{10-character alphanumeric ID; for grouping events}
\item{timestamp}{when the event occurred; uses \link[base:DateTimeClasses]{POSIXct} format}
\item{event}{"serp" or "click"}
\item{position}{position ("ranking") of the clicked search result}
\item{ranking_function}{"A" or "B"}
}
Users in \code{interleaved_data} have no preference, users in
\code{interleaved_data_a} have preference for ranking function "A", and users
in \code{interleaved_data_b} have preference for ranking function "B".}
\usage{
interleaved_bootstraps(sessions, clicks, bootstraps = 1000L)

interleaved_confint(sessions, clicks, bootstraps = 1000L, confidence = 0.95)

interleaved_preference(sessions, clicks)

interleaved_data

interleaved_data_a

interleaved_data_b
}
\arguments{
\item{sessions}{vector of session IDs used to group \code{positions} and
\code{ranking_functions}}

\item{clicks}{vector that shows which ranking function the
clicked search result came from ("A" or "B")}

\item{bootstraps}{number of times to sample unique sessions
(with replacement); 1000 by default}

\item{confidence}{level; 0.95 by default}
}
\description{
Tools for analysis of experiments that use interleaved search
results wherein users receive results from multiple sets of retrieval
functions.
\itemize{
\item \code{interleaved_data} is a fake dataset used for testing and examples;
refer to \strong{Format} section below
\item \code{interleaved_data_a} is a fake dataset used for testing and examples;
"A" is preferred over "B"
\item \code{interleaved_data_b} is a fake dataset used for testing and examples;
"B" is preferred over "A"
\item \code{interleaved_preference} returns a test statistic summarizing the
interleaving experiment; a positive value indicates that A is better
than B, a negative value indicates that B is better than A
\item \code{interleaved_bootstraps} returns a bootstrapped sample of preference
statistics computed by resampling sessions with replacements
\item \code{interleaved_confint} returns a \code{list} with elements "point.est",
"lower", and "upper" (uses \code{interleaved_bootstraps} internally)
\item \code{interleaved_sample_size} estimates the sample size required to detect
a particular effect size with a specified power and significance level
}
}
\examples{
data("interleaved_data", package = "wmf")
x <- interleaved_data[interleaved_data$event == "click", ]
x <- x[order(x$session_id, x$timestamp), ]
data("interleaved_data_a", package = "wmf")
y <- interleaved_data_a[interleaved_data_a$event == "click", ]
y <- y[order(y$session_id, y$timestamp), ]
data("interleaved_data_b", package = "wmf")
z <- interleaved_data_b[interleaved_data_b$event == "click", ]
z <- z[order(z$session_id, z$timestamp), ]

# Bootstrapped preference statistics:

## Data without a clear preference:
b <- interleaved_bootstraps(x$session_id, x$ranking_function)
hist(b)

## Data where A is preferred over B:
b <- interleaved_bootstraps(y$session_id, y$ranking_function)
hist(b)

## Data where B is preferred over A:
b <- interleaved_bootstraps(z$session_id, z$ranking_function)
hist(b)

# Preference statistic confidence intervals:

## Data without a clear preference:
interleaved_confint(x$session_id, x$ranking_function)

## Data where A is preferred over B:
interleaved_confint(y$session_id, y$ranking_function)

## Data where B is preferred over A:
interleaved_confint(z$session_id, z$ranking_function)

# Preference statistic calculation:

## Data without a clear preference:
interleaved_preference(x$session_id, x$ranking_function)

## Data where A is preferred over B:
interleaved_preference(y$session_id, y$ranking_function)

## Data where B is preferred over A:
interleaved_preference(z$session_id, z$ranking_function)
}
\references{
\itemize{
\item Chapelle, O., Joachims, T., Radlinski, F., & Yue, Y. (2012). Large-scale
validation and analysis of interleaved search evaluation.
\emph{ACM Transactions on Information Systems}, \strong{30}(1), 1-41.
\href{https://doi.org/10.1145/2094072.2094078}{doi:10.1145/2094072.2094078}
\item Radlinski, F. and Craswell, N. (2013). \href{https://www.microsoft.com/en-us/research/publication/optimized-interleaving-for-online-retrieval-evaluation/}{Optimized interleaving for online retrieval evaluation}.
\emph{ACM International Conference on Web Search and Data Mining (WSDM)}.
\href{https://doi.org/10.1145/2433396.2433429}{doi:10.1145/2433396.2433429}
}
}
\keyword{datasets}
