% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invert_list}
\alias{invert_list}
\title{Invert list}
\usage{
invert_list(x)
}
\arguments{
\item{x}{A named list}
}
\value{
A list with values as keys and keys as values.
}
\description{
Inverts a (named) list such that the values become the keys and
the keys become the values.
}
\examples{
invert_list(list(x = c(1, 2), y = c(2, 3)))
}
\author{
Mikhail Popov
}
