% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_json}
\alias{parse_json}
\title{Parse character vector of JSON}
\usage{
parse_json(x)
}
\arguments{
\item{x}{character vector of JSON}
}
\description{
This is a shortcut for using \code{\link[purrr:map]{purrr::map()}} with
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}. Useful when a column in a \code{data.frame} has JSON.
\strong{Note}: if the intention is to have the parsed data in a column such as
when parsing in \code{\link[dplyr:mutate]{dplyr::mutate()}}, it is \emph{highly} recommended to switch
to a \code{tibble} via \code{\link[tibble:as_tibble]{tibble::as_tibble()}} \emph{first}.
}
