% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{refine_eventlogs}
\alias{refine_eventlogs}
\title{Refine EventLogging data}
\usage{
refine_eventlogs(el_data, dt_cols = NULL, json_cols = NULL)
}
\arguments{
\item{el_data}{EventLogging data}

\item{dt_cols}{character vector of timestamp and date-time column names to
parse; can also be a named list of parsing functions to apply on a
per-column basis, as \code{\link[lubridate:ymd_hms]{lubridate::ymd_hms()}} is used by default}

\item{json_cols}{character vector of JSON-containing column names that need
to be parsed; can also be a named list of parsing functions to apply on a
per-column basis, as \code{\link[=parse_json]{parse_json()}} is used by default}
}
\value{
A \code{tibble} (see \link[tibble:`tibble-package`]{tibble::\code{tibble-package}} for more info)
}
\description{
Converts date-time and JSON columns, removes "event_" prefix
fom column names, and returns \code{tibble}s.
}
\author{
Mikhail Popov
}
