% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/power.R
\name{sample_size_odds}
\alias{sample_size_odds}
\title{calculate sample size given odds ratio}
\usage{
sample_size_odds(odds_ratio, precision = NULL, p_a = NULL,
  conf_level = 0.95, sample_ratio = 1, visualize = FALSE)
}
\arguments{
\item{odds_ratio}{The expected odds ratio. That is, the ratio of the odds of
the outcome in the test group (b) relative to the control group (a).}

\item{precision}{Relative precision dictates your margin of error. Optional
only if \code{pa} is provided.}

\item{p_a}{Expected prevalence of outcome in the control group (a). Optional
only if \code{precision} is provided.}

\item{conf_level}{Desired confidence level. Defaults to 95\%.}

\item{sample_ratio}{Ratio of test group to control group. 1 is even split.}

\item{visualize}{Whether to plot relative precision or prevalence of outcome
 in the control group vs sample size. Can be used to help make a decision.}
}
\value{
If \code{pa} was not provided, returns a data frame containing
 possible values of \eqn{\pi_a} and the appropriate sample size.
 If \code{pa} was provided, returns a single sample size estimate.
}
\description{
calculate sample size given odds ratio
}

