% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxies.R
\name{set_proxies}
\alias{set_proxies}
\title{Set HTTP and HTTPS proxies}
\usage{
set_proxies()
}
\description{
Sets the HTTP and HTTPS proxies when running R on
Wikimedia machines.
}
\examples{
\dontrun{
# This will fail in the cluster
devtools::install_github("ironholds/urltools")

# This will work
set_proxies()
devtools::install_github("ironholds/urltools")
}
}
