% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{theme_facet}
\alias{theme_facet}
\title{Simple theme for facet-ed ggplots}
\usage{
theme_facet(base_size = 12, base_family = "", border = TRUE,
  clean_xaxis = FALSE, ...)
}
\arguments{
\item{base_size}{font size}

\item{base_family}{font family}

\item{border}{whether to add a border around facets}

\item{clean_xaxis}{whether to remove ticks & labels from x-axis}

\item{...}{additional parameters to pass to \code{theme()}}
}
\description{
A minimal theme that puts the legend at the bottom and puts the
facet labels into gray boxes. The border around those can be disabled.
}
\author{
Mikhail Popov & Chelsy Xie
}
