% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{theme_min}
\alias{theme_min}
\title{Simple theme for ggplots}
\usage{
theme_min(base_size = 12, base_family = "", ...)
}
\arguments{
\item{base_size}{font size}

\item{base_family}{font family}

\item{...}{additional parameters to pass to \code{theme()}}
}
\description{
A minimal theme that puts the legend at the bottom.
}
\author{
Mikhail Popov
}
