% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{timeconverters}
\alias{from_log}
\alias{from_mediawiki}
\alias{timeconverters}
\alias{to_log}
\alias{to_mediawiki}
\title{convert to and from common timestamp formats}
\usage{
from_mediawiki(x)

from_log(x)

to_mediawiki(x)

to_log(x)
}
\arguments{
\item{x}{a vector of timestamps}
}
\description{
convert to and from MediaWiki and request log timestamp formats
}
\examples{
from_mediawiki("20150101010301")
}

