% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{update_shardmap}
\alias{update_shardmap}
\title{Update shard map}
\usage{
update_shardmap(dev = FALSE)
}
\arguments{
\item{dev}{logical flag; if true, will write to inst/extdata; updates the
installed map otherwise}
}
\description{
Fetches the latest DB configuration from WMF's MediaWiki
Configuration for mapping
}
