% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmf.R
\docType{package}
\name{wmf}
\alias{wmf}
\alias{wmf-package}
\title{wmf: R Code for Wikimedia Foundation Internal Usage}
\description{
This package contains functions made for Analysts at Wikimedia
Foundation, but can be used by people outside of the Foundation.
}
