% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{write_conditional}
\alias{write_conditional}
\title{Conditionally write out to a file}
\usage{
write_conditional(x, file)
}
\arguments{
\item{x}{the object to write out}

\item{file}{the path to the file to use.}
}
\description{
if the file already exists, append. If it
doesn't, create!
}
