% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{build_query}
\alias{build_query}
\title{Builds a MySQL query aimed at the EventLogging-centric formats}
\usage{
build_query(fields, table, date = NULL, conditionals = NULL)
}
\arguments{
\item{fields}{the SELECT statement.}

\item{table}{the table to use.}

\item{date}{the date to restrict to. If NULL, yesterday will be used.}

\item{any}{other conditionals to include in the WHERE statement.}
}
\description{
constructs a MySQL query with a conditional around date.
This is aimed at eventlogging, where the date/time is always "timestamp".
}
