% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{connection_details}
\alias{connection_details}
\title{Connection details}
\usage{
connection_details(dbname)
}
\arguments{
\item{dbname}{e.g. "enwiki"; can be a vector of multiple database names}
}
\value{
a named \code{list} of \code{list(host, port)}s
}
\description{
Figure out connection details (host name and port) based on
database name.
}
\references{
\href{https://wikitech.wikimedia.org/wiki/Analytics/Data_access#MariaDB_replicas}{wikitech:Data_access#MariaDB_replicas}
}
