% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mysql.R
\name{mysql}
\alias{mysql}
\alias{mysql_exists}
\alias{mysql_read}
\alias{mysql_write}
\title{query MySQL databases}
\usage{
mysql_read(query, database)

mysql_exists(database, table_name)

mysql_write(x, database, table_name, ...)
}
\arguments{
\item{query}{a SQL query.}

\item{database}{the name of the database to query.}

\item{...}{further arguments to pass to dbWriteTable. See ?dbWriteTable for more details.}

\item{table}{the name of a table to check for the existence of or create, depending on the
function.}
}
\description{
read from, write to, and check data from the MySQL databases and tables in the Wikimedia
cluster. Assumes the presence of a validly formatted .my.cnf.
}
\seealso{
\code{\link{hive_query}} or \code{\link{global_query}}
}

