% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{mysql}
\alias{mysql}
\alias{mysql_connect}
\alias{mysql_read}
\alias{mysql_exists}
\alias{mysql_write}
\alias{mysql_close}
\alias{mysql_disconnect}
\title{Work with MySQL databases}
\usage{
mysql_connect(database, default_file = NULL, hostname = ifelse(database ==
  "log", "db1047.eqiad.wmnet", "analytics-store.eqiad.wmnet"))

mysql_read(query, database, con = NULL)

mysql_exists(database, table_name, con = NULL)

mysql_write(x, database, table_name, con = NULL, ...)

mysql_close(con)

mysql_disconnect(con)
}
\arguments{
\item{database}{name of the database to query}

\item{default_file}{name of a config file containing username and password
to use when connecting}

\item{hostname}{name of the machine to connect to, which depends on whether
\code{query} is used to fetch from the \strong{log} \code{database} (in which case
connect to "db1047.eqiad.wmnet") or a MediaWiki ("content") DB (in which
case connect to "analytics-store.eqiad.wmnet" as before)}

\item{query}{SQL query}

\item{con}{MySQL connection returned by \code{\link[=mysql_connect]{mysql_connect()}}; \emph{optional} -- if
not provided, a temporary connection will be opened up}

\item{table_name}{name of a table to check for the existence of or create,
depending on the function}

\item{x}{a \code{data.frame} to write}

\item{...}{additional arguments to pass to \code{dbWriteTable}}
}
\description{
Read from, write to, and check data from the MySQL databases and
tables in the Wikimedia cluster. Assumes the presence of a validly
formatted configuration file.
}
\seealso{
\code{\link[=query_hive]{query_hive()}} or \code{\link[=global_query]{global_query()}}
}
