% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null2na}
\alias{null2na}
\title{Turns Null Into Character "NA"}
\usage{
null2na(x)
}
\arguments{
\item{x}{A list}
}
\value{
If any element from the input list is NULL, they will be turned into character
 "NA". Otherwise, return the original list.
}
\description{
The function turns NULL in a list into character "NA".
}
