% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hive.R
\name{query_hive}
\alias{query_hive}
\title{Query Hadoop cluster with Hive}
\usage{
query_hive(query, override_jars = FALSE, heap_size = 1024)
}
\arguments{
\item{query}{a Hive query}

\item{override_jars}{A logical flag indicating whether to override the path.
Hive on WMF's analytics machine(s) loads some JARs by default, so if your
query uses an updated version of an existing UDF and you want to load the
JAR that you built yourself, set this to \code{TRUE}. See
\href{https://wikitech.wikimedia.org/wiki/Analytics/Systems/Cluster/Hive/QueryUsingUDF#Testing_changes_to_existing_udf}{Testing changes to existing UDF}
for more details.}

\item{heap_size}{\code{HADOOP_HEAPSIZE}; default is 1024 (alt: 2048 or 4096)}
}
\value{
a \code{data.frame} containing the results of the query, or a \code{TRUE} if
the user has chosen to write straight to file.
}
\description{
Queries Hive
}
\section{escaping}{

\code{hive_query} works by running the query you provide through the CLI via a
\code{\link[=system]{system()}} call. As a result, single escapes for meaningful characters
(such as quotes) within the query will not work: R will interpret them
only as escaping that character /within R/. Double escaping (\) is thus
necessary, in the same way that it is for regular expressions.
}

\section{Handling our hadoop/hive setup}{

The \code{webrequests} table is documented \href{https://wikitech.wikimedia.org/wiki/Analytics/Systems/Cluster/Hive}{on Wikitech},
which also provides \href{https://wikitech.wikimedia.org/wiki/Analytics/Systems/Cluster/Hive/Queries}{a set of example queries}. When it comes to manipulating the rows with Java before they get to you, Nuria has written a
\href{https://wikitech.wikimedia.org/wiki/Analytics/Systems/Cluster/Hive/QueryUsingUDF}{brief tutorial on loading UDFs}
which should help if you want to engage in that.
}

\examples{
\dontrun{
query_hive("USE wmf; DESCRIBE webrequest;")
}
}
\seealso{
\code{\link[lubridate:ymd_hms]{lubridate::ymd_hms()}} for converting the "dt" column in the
webrequests table to proper datetime, and \code{\link[=mysql_read]{mysql_read()}} and
\code{\link[=global_query]{global_query()}} for querying our MySQL databases
}
