% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{timeconverters}
\alias{timeconverters}
\alias{from_mediawiki}
\alias{from_log}
\alias{to_mediawiki}
\alias{to_log}
\title{Convert to and from common timestamp formats}
\usage{
from_mediawiki(x)

from_log(x)

to_mediawiki(x)

to_log(x)
}
\arguments{
\item{x}{a vector of timestamps}
}
\description{
Convert to and from MediaWiki and request log timestamp formats
}
\examples{
from_mediawiki("20150101010301")
}
\seealso{
\code{\link[lubridate:ymd_hms]{lubridate::ymd_hms()}}
}
\author{
Oliver Keyes
}
