% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time.R
\name{timeconverters}
\alias{from_mediawiki}
\alias{from_requestlog}
\alias{timeconverters}
\alias{to_mediawiki}
\alias{to_requestlog}
\title{convert to and from common timestamp formats}
\usage{
from_mediawiki(x)

from_requestlog(x)

to_mediawiki(x)

to_requestlog(x)
}
\arguments{
\item{x}{a vector of timestamps}
}
\description{
convert to and from MediaWiki and request log timestamp formats
}
\examples{
from_mediawiki("20150101010301")
}

