% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmfdata-package.R
\docType{package}
\name{wmfdata-package}
\alias{wmfdata}
\alias{wmfdata-package}
\title{wmfdata: R Tools For Wikimedia Foundation's Analysts And Data Scientists}
\description{
This package contains functions made for Analysts and Scientists at
    the Wikimedia Foundation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gerrit.wikimedia.org/r/plugins/gitiles/wikimedia/discovery/wmf/}
  \item Report bugs at \url{https://phabricator.wikimedia.org/maniphest/task/create/?projects=Product-Analytics&assigned=mpopov}
}

}
\author{
\strong{Maintainer}: Mikhail Popov \email{mikhail@wikimedia.org}

Authors:
\itemize{
  \item Os Keyes (No longer employed at the Foundation)
}

Other contributors:
\itemize{
  \item Chelsy Xie (No longer employed at the Foundation) [contributor]
  \item  Wikimedia Foundation [copyright holder]
}

}
\keyword{internal}
